requirements.txt
modbus_tk==1.1.3
pyserial==3.5
